requirements.txt
beautifulsoup4
requests
selenium
webdriver-manager
pandas
python-dotenv
lxml
